move external call to clients package
