+++ pkg/controller/appsync/appsync.go
@@ -16,13 +16,11 @@
 package appsync
 
 import (
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"net/http"
 
 	"github.com/google/exposure-notifications-server/pkg/logging"
+	"github.com/google/exposure-notifications-verification-server/pkg/clients"
 	"github.com/google/exposure-notifications-verification-server/pkg/config"
 	"github.com/google/exposure-notifications-verification-server/pkg/controller"
 	"github.com/google/exposure-notifications-verification-server/pkg/database"
@@ -58,22 +56,8 @@ func (c *Controller) HandleSync() http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		logger := logging.FromContext(r.Context()).Named("appsync.HandleSync")
 
-		if c.config.AppSyncURL == "" {
-			controller.InternalError(w, r, c.h, errors.New("no APP_SYNC_URL configured"))
-			return
-		}
-
-		client := http.Client{Timeout: c.config.Timeout}
-		resp, err := client.Get(c.config.AppSyncURL)
+		apps, err := clients.AppSync(c.config.AppSyncURL, c.config.Timeout, c.config.FileSizeLimitBytes)
 		if err != nil {
-			controller.InternalError(w, r, c.h, err)
-			return
-		}
-
-		var apps AppsResponse
-		defer resp.Body.Close()
-
-		if err := json.NewDecoder(io.LimitReader(resp.Body, c.config.FileSizeLimitBytes)).Decode(&apps); err != nil {
 			controller.InternalError(w, r, c.h, err)
 			return
 		}
